# openapi/oh.yaml
openapi: 3.1.0
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
- url: http://localhost:8080/openhospital-api-0.1.0
security:
- bearerAuth: []
...
paths:
...
  /plugins:
    get:
      tags:
      - Plugins
      summary: List all installed plugins
      operationId: listPlugins
      responses:
        "200":
          description: Plugin list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PluginDTO"
        "401":
          description: Unauthorized
      security:
      - bearerAuth: []
  /plugins/{pluginId}:
    get:
      tags:
      - Plugins
      summary: Get a single plugin by ID
      operationId: getPlugin
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Plugin found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginDTO"
        "404":
          description: Plugin not found
        "401":
          description: Unauthorized
      security:
      - bearerAuth: []
    delete:
      tags:
      - Plugins
      summary: Uninstall a plugin
      description: "Calls onStop and onUninstall, then removes the plugin record.\
        \ Approval and event history is preserved."
      operationId: uninstallPlugin
      parameters:
      - name: pluginId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Plugin uninstalled
        "404":
          description: Plugin not found
        "401":
          description: Unauthorized
      security:
      - bearerAuth: []
...
  /plugins/{pluginId}/proxy/{connectionKey}/**:
    get:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    put:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_2
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    post:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_1
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    delete:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_3
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    options:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_6
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    head:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_5
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
    patch:
      tags:
      - Plugin Proxy
      summary: Proxy request to the plugin's external system
      description: Forwards the request to the URL configured for the plugin by the
        declared and approved connection key.
      operationId: proxy_4
      parameters:
      - name: pluginId
        in: path
        description: Plugin identifier
        required: true
        schema:
          type: string
      - name: connectionKey
        in: path
        description: Manifest external connection key
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
components:
  schemas:
...
    BundleDescriptorDTO:
      type: object
      properties:
        entry:
          type: string
        remoteName:
          type: string
...
    RouteDescriptorDTO:
      type: object
      properties:
        path:
          type: string
        label:
          type: string
        menuPath:
          type: string
        permission:
          type: string
    SlotContributionDTO:
      type: object
      properties:
        slotId:
          type: string
        mode:
          type: string
        exposedModule:
          type: string
    UiContributionDTO:
      type: object
      properties:
        bundle:
          $ref: "#/components/schemas/BundleDescriptorDTO"
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotContributionDTO"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteDescriptorDTO"
...
    PluginDTO:
      type: object
      description: An installed Open Hospital plugin
      properties:
        pluginId:
          type: string
          description: Plugin identifier in reverse-domain notation
          example: org.isf.plugin.example.patientaudit
        version:
          type: string
          description: Plugin version (semver)
          example: 1.0.0
        name:
          type: string
          description: Human-readable plugin name
          example: Patient Audit Log
        status:
          type: string
          description: Current lifecycle status
          enum:
          - VALIDATING
          - ACTIVE
          - DISABLED
          - FAILED
        installedAt:
          type: string
          format: date-time
          description: When the plugin was installed
        installedBy:
          type: string
          description: Username of the administrator who installed the plugin
        uiContribution:
          $ref: "#/components/schemas/UiContributionDTO"
          description: "UI contribution declared by the plugin - bundle entry, slots with exposedModule, routes"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT